python/src/auth/requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
APScheduler==3.6.3
astor==0.8.1
astroid==2.15.6
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
billiard==4.1.0
blessed==1.20.0
blinker==1.6.2
build==0.10.0
CacheControl==0.13.1
cachetools==4.2.4
celery==5.3.1
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudevents==1.9.0
contourpy==1.1.0
crashtest==0.4.1
cycler==0.11.0
deprecation==2.1.0
dill==0.3.7
distlib==0.3.7
dnspython==2.4.2
dulwich==0.21.6
email-validator==2.0.0.post2
fastapi==0.101.0
filelock==3.12.2
Flask==2.3.3
Flask-Cors==3.0.8
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
functions-framework==3.4.0
gast==0.4.0
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.34
google-api-core==1.16.0
google-auth==1.11.2
google-auth-oauthlib==1.0.0
google-cloud==0.34.0
google-cloud-vision==0.42.0
google-pasta==0.2.0
googleapis-common-protos==1.51.0
grpcio==1.57.0
gunicorn==20.1.0
h11==0.14.0
h5py==3.9.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
idna==2.10
importlib-metadata==6.8.0
imutils==0.5.3
inquirer==3.1.3
installer==0.7.0
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
keras==2.13.1
keyring==24.2.0
kiwisolver==1.4.5
kombu==5.3.1
lazy-object-proxy==1.9.0
libclang==16.0.6
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.1.0
motor==3.2.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mysql-connector-python==8.1.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
open-interpreter==0.1.1
openai==0.27.10
opencv-python==4.8.0.76
opt-einsum==3.3.0
orjson==3.9.5
packaging==23.1
passlib==1.7.4
pexpect==4.8.0
Pillow==10.0.0
pkginfo==1.9.6
platformdirs==3.10.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
prompt-toolkit==3.0.39
protobuf==4.21.12
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
Pygments==2.16.1
PyJWT==2.8.0
pylint==2.17.5
pymongo==4.5.0
PyMySQL==0.9.3
pyparsing==3.0.9
PyPDF2==3.0.1
pyproject_hooks==1.0.0
PyQt3D==5.15.6
PyQt5==5.15.9
PyQt5-sip==12.12.1
PyQt6==6.5.2
PyQt6-3D==6.5.0
PyQt6-Charts==6.5.0
PyQt6-DataVisualization==6.5.0
PyQt6-NetworkAuth==6.5.0
PyQt6-sip==13.5.2
PyQt6-WebEngine==6.5.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-editor==1.0.4
python-engineio==3.11.2
python-multipart==0.0.6
python-socketio==4.4.0
pytz==2019.3
PyYAML @ file:///private/tmp/pyyaml-20230717-6270-1iiwi1j/PyYAML-6.0.1
QScintilla==2.14.1
rapidfuzz==2.15.1
readchar==4.0.5
redis==5.0.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.5.2
rsa==4.0
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.2
sentencepiece==0.1.99
shellingham==1.5.3
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
spotipy==2.23.0
SQLAlchemy==1.3.13
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
TBB==0.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
textbase-client==0.1.3
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
tokentrim==0.1.10
tomlkit==0.12.1
torch==2.0.1
tqdm==4.66.1
transformers==4.31.0
trove-classifiers==2023.8.7
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==1.25.11
uvicorn==0.23.2
uvloop==0.17.0
vine==5.0.0
virtualenv==20.24.4
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.7
wget==3.2
wrapt==1.15.0
xattr==0.10.1
yarl==1.9.2
yaspin==3.0.0
zipp==3.16.2
